+++ src/com/dkhs/portfolio/ui/fragment/FragmentCompare.java
@@ -421,7 +421,7 @@ public class FragmentCompare extends Fragment implements OnClickListener, Fragme
                 }
 
                 float increaseValue = (object.getEnd() - object.getBegin()) / object.getBegin();
-                tvIncreaseValue.setText(StringFromatUtils.get2PointPercent(increaseValue * 100));
+                tvIncreaseValue.setText(StringFromatUtils.get2PointPercent((long) (increaseValue * 100)));
                 if (increaseValue > 0) {
                     increaseView.setBackgroundColor(ColorTemplate.DEF_RED);
                 } else {
